Refactored library name
