--- CTools/Exceptions/__init__.py
@@ -1,7 +1,0 @@
-"""
-CTools custom exceptions.
-"""
-
-from CTools.Exceptions.CircuitToolsExceptions import *
-from CTools.Exceptions.LogicGateExceptions import *
-from CTools.Exceptions.LatchesExceptions import *
