--- CTools/LogicGates/__init__.py
@@ -1,13 +1,0 @@
-"""
-CTools Logic Gates package.
-
-This package includes the  **NOT, AND, NAND, OR, XOR, NOR** logic gates.
-"""
-
-
-from CTools.LogicGates.Not import *
-from CTools.LogicGates.And import *
-from CTools.LogicGates.Nand import *
-from CTools.LogicGates.Or import *
-from CTools.LogicGates.Xor import *
-from CTools.LogicGates.Nor import *
