--- CTools/Multiplexers/__init__.py
@@ -1,10 +1,0 @@
-"""
-PyCircTools Multiplexers package.
-
-This package includes some Multiplexers and Demultiplexers (from 1 to 2 to 1 to 16)
-"""
-
-from CTools.Multiplexers.Mux2to1 import *
-from CTools.Multiplexers.Mux4to1 import *
-from CTools.Multiplexers.Mux8to1 import *
-from CTools.Multiplexers.Mux16to1 import *
