--- CTools/__init__.py
@@ -1,8 +1,0 @@
-"""
-CTools is a python 3 package that provides the user with common circuit elements. It aims to help the implementation of
-circuits using code, serving as an alternative to programs like ModelSIM.
-"""
-
-from CTools.Exceptions import *
-from CTools.LogicGates import *
-from CTools.Latches import *
