+++ PyCircTools/Latches/SRLatch.py
@@ -1,5 +1,5 @@
-from  CTools.Exceptions.CircuitToolsExceptions import NotTruthValue
-from CTools.LogicGates import And, Nor
+from  PyCircTools.Exceptions.CircuitToolsExceptions import NotTruthValue
+from PyCircTools.LogicGates import And, Nor
 
 
 class SRLatch:
