+++ PyCircTools/Latches/__init__.py
@@ -3,3 +3,5 @@ PyCircTools Latches package.
 
 This package includes **** latches.
 """
+
+from PyCircTools.Latches.SRLatch import *
