+++ PyCircTools/LogicGates/And.py
@@ -1,14 +1,14 @@
-from CTools.Exceptions.CircuitToolsExceptions import NotTruthValue
-from CTools.Exceptions.LogicGateExceptions import NonPositiveInput, NotAnInput
+from PyCircTools.Exceptions.CircuitToolsExceptions import NotTruthValue
+from PyCircTools.Exceptions.LogicGateExceptions import NonPositiveInput, NotAnInput
 
 
-class Xor:
+class And:
     """
-    Xor logic gate. Can have any number of inputs greater or equal than 1.
+    And logic gate. Can have any number of inputs greater or equal than 1.
     """
     def __init__(self, inputNumber=2):
         """
-        Xor class constructor method.
+        And class constructor method.
 
         :param inputNumber: Number of inputs for the gate. Defaults to two.
         :type inputNumber: int
@@ -75,13 +75,24 @@ class Xor:
         self.__calculate_output()
         return self
 
+    @staticmethod
+    def __create_input(number):
+        return [False] * number
+
+    def __calculate_output(self):
+        output = True
+        for value in self.input:
+            output = output and value
+        self.output = output
+        return self
+
     def add_input(self):
         """
         Method add_input adds a new input to a logic gate.
         """
         self.input.append(False)
         self.numOfInputs += 1
-        self.__calculate_output()
+        self.output = False
         return self
 
     def remove_input(self):
@@ -92,14 +103,3 @@ class Xor:
         self.numOfInputs -= 1
         self.__calculate_output()
         return self
-
-    def __calculate_output(self):
-        output = False
-        for value in self.input:
-            output = ((output and not value) or (not output and value))
-        self.output = output
-        return self
-
-    @staticmethod
-    def __create_input(number):
-        return [False] * number
