+++ PyCircTools/LogicGates/Nor.py
@@ -1,14 +1,14 @@
-from CTools.Exceptions.CircuitToolsExceptions import NotTruthValue
-from CTools.Exceptions.LogicGateExceptions import NonPositiveInput, NotAnInput
+from PyCircTools.Exceptions.CircuitToolsExceptions import NotTruthValue
+from PyCircTools.Exceptions.LogicGateExceptions import NonPositiveInput, NotAnInput
 
 
-class And:
+class Nor:
     """
-    And logic gate. Can have any number of inputs greater or equal than 1.
+    Nor logic gate. Can have any number of inputs greater or equal than 1.
     """
     def __init__(self, inputNumber=2):
         """
-        And class constructor method.
+        Norclass constructor method.
 
         :param inputNumber: Number of inputs for the gate. Defaults to two.
         :type inputNumber: int
@@ -75,24 +75,13 @@ class And:
         self.__calculate_output()
         return self
 
-    @staticmethod
-    def __create_input(number):
-        return [False] * number
-
-    def __calculate_output(self):
-        output = True
-        for value in self.input:
-            output = output and value
-        self.output = output
-        return self
-
     def add_input(self):
         """
         Method add_input adds a new input to a logic gate.
         """
         self.input.append(False)
         self.numOfInputs += 1
-        self.output = False
+        self.__calculate_output()
         return self
 
     def remove_input(self):
@@ -103,3 +92,14 @@ class And:
         self.numOfInputs -= 1
         self.__calculate_output()
         return self
+
+    def __calculate_output(self):
+        output = False
+        for value in self.input:
+            output = output or value
+        self.output = not output
+        return self
+
+    @staticmethod
+    def __create_input(number):
+        return [False] * number
