+++ PyCircTools/LogicGates/Not.py
@@ -1,4 +1,4 @@
-from CTools.Exceptions.CircuitToolsExceptions import NotTruthValue
+from PyCircTools.Exceptions.CircuitToolsExceptions import NotTruthValue
 
 
 class Not:
