+++ PyCircTools/LogicGates/Or.py
@@ -1,14 +1,14 @@
-from CTools.Exceptions.CircuitToolsExceptions import NotTruthValue
-from CTools.Exceptions.LogicGateExceptions import NonPositiveInput, NotAnInput
+from PyCircTools.Exceptions.CircuitToolsExceptions import NotTruthValue
+from PyCircTools.Exceptions.LogicGateExceptions import NonPositiveInput, NotAnInput
 
 
-class Nand:
+class Or:
     """
-    Nand logic gate. Can have any number of inputs greater or equal than 1.
+    Or logic gate. Can have any number of inputs greater or equal than 1.
     """
     def __init__(self, inputNumber=2):
         """
-        Nand class constructor method.
+        Or class constructor method.
 
         :param inputNumber: Number of inputs for the gate. Defaults to two.
         :type inputNumber: int
@@ -75,13 +75,6 @@ class Nand:
         self.__calculate_output()
         return self
 
-    def __calculate_output(self):
-        output = True
-        for value in self.input:
-            output = output and value
-        self.output = not output
-        return self
-
     def add_input(self):
         """
         Method add_input adds a new input to a logic gate.
@@ -96,10 +89,17 @@ class Nand:
         Removes the last input of a logic gate.
         """
         self.input.pop()
-        self.__calculate_output()
         self.numOfInputs -= 1
+        self.__calculate_output()
+        return self
+
+    def __calculate_output(self):
+        output = False
+        for value in self.input:
+            output = output or value
+        self.output = output
         return self
 
     @staticmethod
-    def __create_input(num):
-        return [False] * num
+    def __create_input(number):
+        return [False] * number
