+++ PyCircTools/Multiplexers/Mux8to1.py
@@ -1,24 +1,24 @@
-from CTools.Exceptions.MultiplexersExceptions import NonExistingInput, NonExistingControlSignal
-from CTools.Exceptions.CircuitToolsExceptions import NotTruthValue
+from PyCircTools.Exceptions.MultiplexersExceptions import NonExistingInput, NonExistingControlSignal
+from PyCircTools.Exceptions.CircuitToolsExceptions import NotTruthValue
 
 
-class Mux4to1:
+class Mux8to1:
     """
-    4 to 1 Multiplexer. Takes two bits as a Set input.
+    8 to 1 Multiplexer. Takes three bits as Set input.
     """
     def __init__(self):
         """
-        Mux4to1 class constructor method.
+        Mux8to1 class constructor method.
         """
-        self.input = [False] * 4
-        self.set = [False] * 2
+        self.input = [False] * 8
+        self.set = [False] * 3
         self.output = self.__calculate_output()
 
     def get_input(self, num):
         """
         Method get_input is used to get the value of input[num]
 
-        :param num: Number of the input you want to get. It is a number from 0 to 3.
+        :param num: Number of the input you want to get. It is a number from 0 to 7.
         :type num: int
         :raises NonExistingInput: Raised when a Multiplexer/Demultiplexer doesn't have the input asked for.
         :return: Value of the desired input (num).
@@ -33,7 +33,7 @@ class Mux4to1:
         """
         Method get_set gets the value of the set[setNum] control signal.
 
-        :param setNum: Number of the set control signal you want to get, either 0 or 1.
+        :param setNum: Number of the set control signal you want to get, ranges from 0 to 3.
         :type setNum: int
         :raises NonExistingControlSignal: Raised when a Multiplexer/Demultiplexer doesn't have setNum control signal.
         :return: Value of the desired set control signal (setNum).
@@ -46,7 +46,7 @@ class Mux4to1:
 
     def get_output(self):
         """
-        Method get_output gets the output of the 4 to 1 Multiplexer.
+        Method get_output gets the output of the 8 to 1 Multiplexer.
 
         :return: Value of the output.
         :rtype: bool
@@ -97,13 +97,18 @@ class Mux4to1:
 
     def __calculate_output(self):
         """
-        Private method __calculate_output calculates the output signal using the 4-to-1 mux formula.
+        Private method __calculate_output calculates the output signal using the 8-to-1 mux formula.
 
         :return: Value of the output signal.
         :rtype: bool
         """
-        first_minterm = not self.set[0] and not self.set[1] and self.input[0]
-        second_minterm = self.set[0] and not self.set[1] and self.input[1]
-        third_minterm = not self.set[0] and self.set[1] and self.input[2]
-        fourth_minterm = self.set[0] and self.set[1] and self.input[3]
-        return first_minterm or second_minterm or third_minterm or fourth_minterm
+        first = not self.set[2] and not self.set[1] and not self.set[0] and self.input[0]
+        second = not self.set[2] and not self.set[1] and self.set[0] and self.input[1]
+        third = not self.set[2] and self.set[1] and not self.set[0] and self.input[2]
+        fourth = not self.set[2] and self.set[1] and self.set[0] and self.input[3]
+        fifth = self.set[2] and not self.set[1] and not self.set[0] and self.input[4]
+        sixth = self.set[2] and not self.set[1] and self.set[0] and self.input[5]
+        seventh = self.set[2] and self.set[1] and not self.set[0] and self.input[6]
+        eight = self.set[2] and self.set[1] and self.set[0] and self.input[7]
+
+        return first or second or third or fourth or fifth or sixth or seventh or eight
