+++ PyCircTools/Multiplexers/__init__.py
@@ -1,0 +1,10 @@
+"""
+PyCircTools Multiplexers package.
+
+This package includes some Multiplexers and Demultiplexers (from 1 to 2 to 1 to 16)
+"""
+
+from PyCircTools.Multiplexers.Mux2to1 import *
+from PyCircTools.Multiplexers.Mux4to1 import *
+from PyCircTools.Multiplexers.Mux8to1 import *
+from PyCircTools.Multiplexers.Mux16to1 import *
