+++ tests/Multiplexers_tests/test_Mux16to1.py
@@ -1,7 +1,7 @@
 import itertools
 
-from CTools import NotTruthValue, NonExistingControlSignal, NonExistingInput
-from CTools.Multiplexers import Mux16to1
+from PyCircTools import NotTruthValue, NonExistingControlSignal, NonExistingInput
+from PyCircTools.Multiplexers import Mux16to1
 
 
 def generate_inputs(num):
